requirements.txt
pandas>=2.0
numpy>=1.26
scipy>=1.13
scikit-learn>=1.4
requests>=2.31
openpyxl>=3.1
matplotlib>=3.8
seaborn>=0.13
statsmodels>=0.14
openai>=1.30        # only needed if LLM_BACKEND="openai"
tqdm>=4.66
python-dotenv>=1.0

# ── Funding Alert Agent ────────────────────────────────────────────────────────
feedparser>=6.0         # RSS/Atom feed parsing (Gates, Wellcome)
beautifulsoup4>=4.12    # HTML scraping (WHO page)
lxml>=4.9               # fast HTML/XML parser backend for BeautifulSoup
apscheduler>=3.10       # daily scheduler for run_agent.py
